Cleanup - moved constants and interface files
