+++ ui/vic-ui-h5c/vic/src/vic-webapp/src/app/container-view/container.model.ts
@@ -15,66 +15,66 @@
 */
 
 import {
-    VirtualMachine,
-    ContainerVmResponse
-} from '../vm.interface';
+  VirtualMachine,
+  ContainerVmResponse
+} from '../interfaces/vm.interface';
 
 const FORWARD_SLASH = '/';
 const COLON = ':';
 
 export class ContainerVm implements VirtualMachine {
-    public parentObj: {
-        id: string;
-        value: string;
-        type: string;
-    };
-    public vmId: string;
-    public readonly isVCH: boolean = false;
-    public readonly isContainer: boolean = true;
-    public name: string;
-    public overallStatus: string;
-    public powerState: string;
-    public containerName: string;
-    public parentObjectName: string;
-    public imageName: string;
-    public portMapping: string;
-    public overallCpuUsage: number;
-    public guestMemoryUsage: number;
-    public committedStorage: number;
-    public resourceConfig: any;
+  public parentObj: {
+    id: string;
+    value: string;
+    type: string;
+  };
+  public vmId: string;
+  public readonly isVCH: boolean = false;
+  public readonly isContainer: boolean = true;
+  public name: string;
+  public overallStatus: string;
+  public powerState: string;
+  public containerName: string;
+  public parentObjectName: string;
+  public imageName: string;
+  public portMapping: string;
+  public overallCpuUsage: number;
+  public guestMemoryUsage: number;
+  public committedStorage: number;
+  public resourceConfig: any;
 
-    constructor(data: ContainerVmResponse) {
-        try {
-            // populate vm information
-            const splitVmId = data.id.split(FORWARD_SLASH);
-            this.parentObj = {
-                id: `urn:vmomi:${data.resourcePool.type}:${data.resourcePool.value}${COLON}${splitVmId[0]}`,
-                type: data.resourcePool.type,
-                value: data.resourcePool.value
-            };
-            this.parentObjectName = data.parentObjectName;
-            this.vmId = `urn:vmomi:VirtualMachine:${splitVmId[1]}${COLON}${splitVmId[0]}`;
-            this.name = data.name;
-            this.overallStatus = data.overallStatus;
-            this.powerState = data.powerState;
-            this.containerName = data.containerName;
-            this.imageName = data.imageName;
-            this.portMapping = data.portMapping;
-            this.overallCpuUsage = data.overallCpuUsage;
-            this.guestMemoryUsage = data.guestMemoryUsage;
-            this.committedStorage = data.committedStorage;
-            this.resourceConfig = data.resourceConfig;
-        } catch (e) {
-            throw new Error('response does not fit into the required type! ' +
-                e.toString());
-        }
+  constructor(data: ContainerVmResponse) {
+    try {
+      // populate vm information
+      const splitVmId = data.id.split(FORWARD_SLASH);
+      this.parentObj = {
+        id: `urn:vmomi:${data.resourcePool.type}:${data.resourcePool.value}${COLON}${splitVmId[0]}`,
+        type: data.resourcePool.type,
+        value: data.resourcePool.value
+      };
+      this.parentObjectName = data.parentObjectName;
+      this.vmId = `urn:vmomi:VirtualMachine:${splitVmId[1]}${COLON}${splitVmId[0]}`;
+      this.name = data.name;
+      this.overallStatus = data.overallStatus;
+      this.powerState = data.powerState;
+      this.containerName = data.containerName;
+      this.imageName = data.imageName;
+      this.portMapping = data.portMapping;
+      this.overallCpuUsage = data.overallCpuUsage;
+      this.guestMemoryUsage = data.guestMemoryUsage;
+      this.committedStorage = data.committedStorage;
+      this.resourceConfig = data.resourceConfig;
+    } catch (e) {
+      throw new Error('response does not fit into the required type! ' +
+        e.toString());
     }
+  }
 
-    get parentType() {
-        return this.parentObj.type;
-    }
+  get parentType() {
+    return this.parentObj.type;
+  }
 
-    get id(): string {
-        return this.vmId;
-    }
+  get id(): string {
+    return this.vmId;
+  }
 }
