+++ ui/vic-ui-h5c/vic/src/vic-webapp/src/app/interfaces/vm.interface.ts
