+++ ui/vic-ui-h5c/vic/src/vic-webapp/src/app/shared/constants/vm.constants.ts
