+++ ui/vic-ui-h5c/vic/src/vic-webapp/src/app/summary-portlet/vch-portlet/vch-portlet.component.ts
@@ -17,17 +17,17 @@ import { I18nService } from '../../shared/i18n.service';
 import { GlobalsService } from '../../shared/globals.service';
 
 import { Component, Input, OnInit } from '@angular/core';
-import { VirtualMachine } from '../../vm.interface';
+import { VirtualMachine } from '../../interfaces/vm.interface';
 
 import {
-    DEFAULT_VCH_DOCKER_ENDPOINT_LABEL,
-    DEFAULT_VCH_VCH_ADMIN_PORTAL_LABEL
+  DEFAULT_VCH_DOCKER_ENDPOINT_LABEL,
+  DEFAULT_VCH_VCH_ADMIN_PORTAL_LABEL
 } from '../../shared/constants/index';
 
 @Component({
-    selector: 'vic-vch-portlet',
-    styleUrls: ['vch-portlet.scss'],
-    template: `
+  selector: 'vic-vch-portlet',
+  styleUrls: ['vch-portlet.scss'],
+  template: `
     <table class="table table-vertical">
         <tbody>
             <tr>
@@ -50,32 +50,32 @@ import {
     `
 })
 export class VchPortletComponent implements OnInit {
-    @Input() activeVm: VirtualMachine;
+  @Input() activeVm: VirtualMachine;
 
-    public dockerApiEndpointLabel: string;
-    public vchAdminPortalLabel: string;
+  public dockerApiEndpointLabel: string;
+  public vchAdminPortalLabel: string;
 
-    constructor(
-        private i18n: I18nService,
-        private globalsService: GlobalsService
-    ) { }
+  constructor(
+    private i18n: I18nService,
+    private globalsService: GlobalsService
+  ) { }
 
-    /**
-     * Evaluates if the application is running in the vSphere Client environment,
-     * which is called the "Plugin Mode"
-     * @returns true if plugin mode. false if not
-     */
-    isPluginMode() {
-        return this.globalsService.isPluginMode() &&
-            !(<any>this.activeVm)['is_testing'];
-    }
+  /**
+   * Evaluates if the application is running in the vSphere Client environment,
+   * which is called the "Plugin Mode"
+   * @returns true if plugin mode. false if not
+   */
+  isPluginMode() {
+    return this.globalsService.isPluginMode() &&
+      !(<any>this.activeVm)['is_testing'];
+  }
 
-    ngOnInit() {
-        this.dockerApiEndpointLabel = this.isPluginMode() ?
-            this.i18n.translate('vch.dockerApiEndpoint.label') :
-            DEFAULT_VCH_DOCKER_ENDPOINT_LABEL;
-        this.vchAdminPortalLabel = this.isPluginMode() ?
-            this.i18n.translate('vch.vchAdminPortal.label') :
-            DEFAULT_VCH_VCH_ADMIN_PORTAL_LABEL;
-    }
+  ngOnInit() {
+    this.dockerApiEndpointLabel = this.isPluginMode() ?
+      this.i18n.translate('vch.dockerApiEndpoint.label') :
+      DEFAULT_VCH_DOCKER_ENDPOINT_LABEL;
+    this.vchAdminPortalLabel = this.isPluginMode() ?
+      this.i18n.translate('vch.vchAdminPortal.label') :
+      DEFAULT_VCH_VCH_ADMIN_PORTAL_LABEL;
+  }
 }
